requirements.txt
dj-database-url==2.1.0
Django==5.0.1
gunicorn==21.2.0
psycopg2==3.1.17
whitenoise==6.6.0
djangorestframework==3.14.0
django-environ==0.11.2
django_extensions==3.2.3